requirements.txt
numpy==1.23.1
pandas==1.4.3
filetype==1.0.13
opencv-python==4.6.0.66
Pillow==9.2.0
pytest==7.1.2
PyYAML==6.0
scikit-learn==1.1.1
torch==1.12.0
torchvision==0.12.0